fix: chat window not sending all text (#23)
As an example, quest 6.1 requires a phrase that sends a string that
fills the entire window. The arrow key doesn't move all the way to the
right as it's miscalculating how many times to arrow over.

The problem is due to us needing to arrow over 60 times (20 characters x
3 cursor moves per character) - the chat buffer has a problem just
holding down the arrow key for 60 blank spaces. If we type our
characters into the buffer one at a time, moving the cursor as we insert
into the buffer, the cursor doesn't go outside of the chat window and
then properly processes the string.